project-management-api/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
colorlog==6.9.0
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
lazy-model==0.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.2
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.37.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
